--- module12/src/main/java/ru/edu/module12/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package ru.edu.module12.repository;
-
-import ru.edu.module12.entity.SberUser;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository <SberUser, Long> {
-
-}
