--- module12/src/main/java/ru/edu/module12/service/UserInfo.java
@@ -1,47 +1,0 @@
-package ru.edu.module12.service;
-
-public class UserInfo {
-    private String login;
-    private String password;
-    private String name;
-    private String role;
-
-    public UserInfo(String login, String password, String name, String role) {
-        this.login = login;
-        this.password = password;
-        this.name = name;
-        this.role = role;
-    }
-
-    public String getLogin() {
-        return login;
-    }
-
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getRole() {
-        return role;
-    }
-
-    public void setRole(String role) {
-        this.role = role;
-    }
-}
