+++ module12/src/main/java/ru/edu/module13/Application.java
@@ -1,0 +1,13 @@
+package ru.edu.module13;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+
+}
