+++ module12/src/main/java/ru/edu/module13/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package ru.edu.module13.repository;
+
+import ru.edu.module13.entity.SberUser;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository <SberUser, Long> {
+
+}
