+++ module13/src/main/java/ru/edu/module13/entity/SberUser.java
@@ -1,0 +1,37 @@
+package ru.edu.module13.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@Table(name = "sberusers")
+@Data
+public class SberUser {
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    private Long id;
+    private String name;
+    private Integer age;
+
+    public SberUser() {
+    }
+
+    public SberUser(Long id, String name, Integer age) {
+        this.id = id;
+        this.name = name;
+        this.age = age;
+    }
+
+    @Override
+    public String toString() {
+        return "SberUser{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", age=" + age +
+                '}';
+    }
+}
