+++ module13/src/main/java/ru/edu/module13/service/UserService.java
@@ -1,0 +1,45 @@
+package ru.edu.module13.service;
+
+import ru.edu.module13.entity.SberUser;
+import ru.edu.module13.repository.UserRepository;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
+
+@Service
+public class
+UserService {
+    private final UserRepository userRepository;
+
+    public UserService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    public SberUser createUser(SberUser sberUser){
+        return userRepository.save(sberUser);
+    }
+
+    public List<SberUser> getUsers() {
+        return userRepository.findAll();
+    }
+
+    public SberUser updateUser(SberUser user, Long id) {
+        Optional<SberUser> oldUser = userRepository.findById(id);
+        if (oldUser.isPresent()){
+            SberUser updateUser = oldUser.get();
+            updateUser.setName(user.getName());
+            updateUser.setAge(user.getAge());
+            return userRepository.save(updateUser);
+        }else {
+            throw new RuntimeException("User not find!");
+        }
+
+    }
+
+    public void deleteUser(Long id) {
+        userRepository.deleteById(id);
+    }
+
+
+}
