--- module13/src/main/java/ru/sberbank/edu/App.java
@@ -1,13 +1,0 @@
-package ru.sberbank.edu;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
