+++ module13/src/main/resources/application.yml
@@ -1,0 +1,24 @@
+server:
+  port: 8080
+  address: 0.0.0.0
+
+spring:
+  datasource:
+    url: jdbc:postgresql://localhost:5432/postgressber
+    driverClassName: org.postgresql.Driver
+    username: postgres
+    password: 1234
+
+  jpa:
+    defer-datasource-initialization: true
+    hibernate:
+      ddl-auto: create-drop
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.PostgreSQLDialect
+        show_sql: true
+        format_sql: true
+
+  service:
+    description: SberUser service
+    version: 0.1
